@@ -1,31 +1,35 @@
 {
-  "_from": "sqlstring@^2.3.2",
-  "_id": "sqlstring@2.3.2",
+  "_args": [
+    [
+      "sqlstring@2.3.3",
+      "/Users/miareinshagen/Desktop/PROJET_7/PROJET_7_MIA/backend"
+    ]
+  ],
+  "_from": "sqlstring@2.3.3",
+  "_id": "sqlstring@2.3.3",
   "_inBundle": false,
-  "_integrity": "sha512-vF4ZbYdKS8OnoJAWBmMxCQDkiEBkGQYU7UZPtL8flbDRSNkhaXvRJ279ZtI6M+zDaQovVU4tuRgzK5fVhvFAhg==",
+  "_integrity": "sha512-qC9iz2FlN7DQl3+wjwn3802RTyjCx7sDvfQEXchwa6CWOx07/WVfh91gBmQ9fahw8snwGEWU3xGzOt4tFyHLxg==",
   "_location": "/sqlstring",
   "_phantomChildren": {},
   "_requested": {
-    "type": "range",
+    "type": "version",
     "registry": true,
-    "raw": "sqlstring@^2.3.2",
+    "raw": "sqlstring@2.3.3",
     "name": "sqlstring",
     "escapedName": "sqlstring",
-    "rawSpec": "^2.3.2",
+    "rawSpec": "2.3.3",
     "saveSpec": null,
-    "fetchSpec": "^2.3.2"
+    "fetchSpec": "2.3.3"
   },
   "_requiredBy": [
     "/mysql2"
   ],
-  "_resolved": "https://registry.npmjs.org/sqlstring/-/sqlstring-2.3.2.tgz",
-  "_shasum": "cdae7169389a1375b18e885f2e60b3e460809514",
-  "_spec": "sqlstring@^2.3.2",
-  "_where": "/Users/miareinshagen/Desktop/PROJET_7/PROJET_7_MIA/backend/node_modules/mysql2",
+  "_resolved": "https://registry.npmjs.org/sqlstring/-/sqlstring-2.3.3.tgz",
+  "_spec": "2.3.3",
+  "_where": "/Users/miareinshagen/Desktop/PROJET_7/PROJET_7_MIA/backend",
   "bugs": {
     "url": "https://github.com/mysqljs/sqlstring/issues"
   },
-  "bundleDependencies": false,
   "contributors": [
     {
       "name": "Adri Van Houdt",
@@ -53,14 +57,13 @@
       "email": "seregpie@gmail.com"
     }
   ],
-  "deprecated": false,
   "description": "Simple SQL escape and format for MySQL",
   "devDependencies": {
     "beautify-benchmark": "0.2.4",
     "benchmark": "2.1.4",
-    "eslint": "5.16.0",
-    "eslint-plugin-markdown": "1.0.2",
-    "nyc": "15.0.0",
+    "eslint": "7.32.0",
+    "eslint-plugin-markdown": "2.2.1",
+    "nyc": "15.1.0",
     "urun": "0.0.8",
     "utest": "0.0.8"
   },
@@ -89,10 +92,10 @@
   },
   "scripts": {
     "bench": "node benchmark/index.js",
-    "lint": "eslint --plugin markdown --ext js,md .",
+    "lint": "eslint .",
     "test": "node test/run.js",
-    "test-ci": "nyc --reporter=text npm test",
+    "test-ci": "nyc --reporter=lcovonly --reporter=text npm test",
     "test-cov": "nyc --reporter=html --reporter=text npm test"
   },
-  "version": "2.3.2"
+  "version": "2.3.3"
 }
