@@ -2,4 +2,4 @@ module.exports = (sequelize, DataTypes) => {
     const Likes = sequelize.define("Likes")
   
     return Likes;
-  };{}
+  };
