@@ -1,7 +1,8 @@
-import React from 'react';
+import React,{useState,useEffect} from 'react';
 import GlobalStyle from './globalStyles';
-import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
-
+import { BrowserRouter as Router, Switch, Route,Redirect } from 'react-router-dom';
+import { GlobalContext } from './Context/globalContext';
+import axios from 'axios';
 //Pages
 
 import Home from './pages/Home.pages';
@@ -10,21 +11,68 @@ import SignIn from './pages/Signin.pages';
 import Profil from './pages/Profil.pages';
 import Navbar from './components/Navbar/Navbar';
 import Footer from './components/Footer/Footer';
+import PostCreate from './pages/PostCreate.pages';
+
+
 
 function App() {
+
+	const [appContext, setAppContext] = useState({
+    authState : {
+                  username: "",
+                  userid: 0,
+                  isconnected: false,
+                  isadmin:false,
+                  imgprofil: ""
+                },
+    postsState:[]
+  });
+
+    useEffect(() => {
+    axios
+      .get("http://localhost:3500/api/auth/authuser", {
+        headers: {
+          connectedToken: localStorage.getItem("connectedToken"),
+        },
+      })
+      .then((response) => { 
+		  console.log(response)
+        if (response.data.error) {
+          setAppContext({ 
+            ...appContext,
+            authState:{...appContext.authState, isconnected: false }});
+        } else {
+          setAppContext({
+            ...appContext,
+            authState:{
+            username: response.data.username,
+            userid: response.data.id,
+            isconnected: true,
+            isadmin: response.data.isAdminAccount,
+			      imgprofil: response.data.name
+            }
+          });
+        }
+      });
+  }, [appContext.authState.isconnected]);
+
+
   return (
+	   <GlobalContext.Provider value={{ appContext, setAppContext}}>
 		<Router>
 			<GlobalStyle />
 			<Navbar/>
 			<Switch>
 				<Route path="/" exact component={Home} />
-				<Route path="/signup" exact component={SignUp} />
-				<Route path="/signin" exact component={SignIn} />
-        <Route path="/profil" exact component={Profil} />
+				 <Route path="/signup" exact={true} component={()=> {return !appContext.authState.isconnected ? <SignUp/> : < Redirect to="/" />}}/>
+                <Route path="/signin" exact={true} component={()=> {return !appContext.authState.isconnected ? <SignIn/> : < Redirect to="/" />}}/>
+                <Route path="/profil" exact={true} component={()=> {return appContext.authState.isconnected ? <Profil/> : < Redirect to="/" />}}/>
+                <Route path="/createpost" exact={true} component={()=> {return appContext.authState.isconnected ? <PostCreate/> : < Redirect to="/" />}}/>
 			</Switch>
       <Footer />
 			
 		</Router>
+	</GlobalContext.Provider>
 	);
 }
 
