@@ -5,10 +5,10 @@ import {
 	TextWrapper,
 	TopLine,
 	Heading,
+	ContentButton,
 	Subtitle,
 	ImgWrapper,
 	Img,
-	Icons,
 	ContentColumn,
 } from './ContentStyles.js';
 
@@ -20,6 +20,7 @@ export const Content = ({
 	topLine,
 	headline,
 	description,
+	buttonLabel,
 	img,
 	alt,
 	inverse,
@@ -56,7 +57,7 @@ export const Content = ({
 								initial={initial}
 								transition={{ delay: 0.5, duration: 0.6 }}
 								animate={animation}
-								inverse={inverse}
+								inverse={inverse.toString()}
 							>
 								{headline}
 							</Heading>
@@ -64,17 +65,21 @@ export const Content = ({
 								initial={initial}
 								transition={{ delay: 0.7, duration: 0.6 }}
 								animate={animation}
-								inverse={inverse}
+								inverse={inverse.toString()}
 							>
 								{description}
 							</Subtitle>
+							<ContentButton
+								initial={initial}
+								transition={{ delay: 1, duration: 0.6 }}
+								animate={animation}
+								inverse={inverse.toString()}
+								primary={primary}
+							>
+								{buttonLabel}
 
-
+							</ContentButton>
 						</TextWrapper>
-						<Icons>
-							<i class="fas fa-heart"></i>
-							<i class="fas fa-comment"></i>
-						</Icons>
 					</ContentColumn>
 					<ContentColumn
 						initial={initial}
